front-chat/requirements.txt
streamlit
requests
python-dotenv
plotly
google-auth-oauthlib
google-api-python-client
PyJWT>=2.0.0
extra-streamlit-components
